+++ exercicios/para-sala/index.js
@@ -33,22 +33,5 @@ function sum(num1, num2) {
   }
    
 
-class Calculadora {
-    soma = (num1, num2) => num1+num2
-    subtrai = (num1, num2) => num1-num2
-    multiplica = (num1, num2) => num1*num2
-    divide = (num1, num2) => num1/num2
-    fatorial = (num1) => {
-        if(num1 === 0) {
-            return 1
-        }
-        return num1 * this.fatorial(num1 - 1)
-    }
-    somaDivide = (num1, num2, num3) = {
-        const soma = this.soma(num1)
-        
-     }
-}
-    
+ 
 module.exports = {convertCelsiusToFahrenheit,checkEvenNumber,sum,revert,};
-module.exports = {Calculadora};
